+++ app/controllers/projects_controller.rb
@@ -212,6 +212,7 @@ class ProjectsController < ApplicationController
         :name, :description, :acrostic_enabled, :subject_code_name,
         :show_contacts, :show_documents, :show_posts, :disable_all_emails,
         :collect_email_on_surveys, :lockable, :hide_values_on_pdfs,
+        :double_data_entry,
         # Uploaded Logo
         :logo, :logo_uploaded_at, :logo_cache, :remove_logo,
         # Will automatically generate a site if the project has no site