+++ db/migrate/20141114173932_add_double_data_entry_to_projects.rb
@@ -1,0 +1,5 @@
+class AddDoubleDataEntryToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :double_data_entry, :boolean, null: false, default: false
+  end
+end